+++ src/api/objects/toot.ts
@@ -2,7 +2,9 @@
  * Ideally this would be a formal class but for now it's just some helper functions
  * for dealing with Toot objects.
  */
+import LanguageDetect from 'languagedetect';
 import { capitalCase } from "change-case";
+import { detectAll } from 'tinyld';
 import { mastodon } from "masto";
 import { Type } from 'class-transformer';
 const escape = require('regexp.escape');
@@ -15,6 +17,7 @@ import UserData from "../user_data";
 import { ageInSeconds, ageString, timelineCutoffAt, toISOFormat } from "../../helpers/time_helpers";
 import { batchMap, groupBy, sortObjsByProps, sumArray, uniquify, uniquifyByProp } from "../../helpers/collection_helpers";
 import { Config } from "../../config";
+import { FOREIGN_SCRIPTS, LANGUAGE_CODES } from "../../helpers/language_helper";
 import { logTootRemoval, traceLog } from '../../helpers/log_helpers';
 import { repairTag } from "./tag";
 import {
@@ -26,6 +29,10 @@ import {
     htmlToText,
     isImage,
     isVideo,
+    removeEmojis,
+    removeLinks,
+    removeMentions,
+    removeTags,
     replaceEmojiShortcodesWithImageTags,
     replaceHttpsLinks
 } from "../../helpers/string_helpers";
@@ -51,9 +58,21 @@ enum TootVisibility {
 };
 
 const MAX_ID_IDX = 2;
+const LANGUAGE_DETECTOR = new LanguageDetect();
 const MAX_CONTENT_PREVIEW_CHARS = 110;
+const MIN_CHARS_FOR_LANG_DETECT = 8;
+const MIN_ACCURACY_FOR_LANG_REPLACE = 0.4;
+const MIN_VERY_HIGH_ACCURACY_FOR_LANG_REPLACE = 0.9;
 const UNKNOWN = "unknown";
 const BLUESKY_BRIDGY = 'bsky.brid.gy';
+const REPAIR_TOOT = bracketed("repairToot");
+
+const IGNORE_LANGUAGES = [
+    "ber",  // Berber
+    "eo",   // Esperanto
+    "tk",   // Turkmen
+    "tlh",  // Klingon
+];
 
 const PROPS_THAT_CHANGE: (keyof Toot)[] = [
     "favouritesCount",
@@ -95,6 +114,9 @@ interface TootObj extends SerializableToot {
     ageInHours: () => number;
     containsString: (str: string) => boolean;
     containsTag: (tag: string | MastodonTag, fullScan?: boolean) => boolean;
+    contentString: () => string;
+    contentShortened: (maxChars?: number) => string;
+    contentWithEmojis: (fontSize?: number) => string;
     describe: () => string;
     homeserverURL: () => Promise<string>;
     isDM: () => boolean;
@@ -276,10 +298,15 @@ export default class Toot implements TootObj {
         return this.mentions.some((mention) => mention.acct == MastoApi.instance.user.webfingerURI);
     }
 
+    // Return the content field stripped of HTML tags
+    contentString(): string {
+        return htmlToText(this.reblog?.content || this.content || "");
+    }
+
     // Shortened string of content property stripped of HTML tags
     contentShortened(maxChars?: number): string {
         maxChars ||= MAX_CONTENT_PREVIEW_CHARS;
-        let content = htmlToText(this.reblog?.content || this.content || "");
+        let content = this.contentString();
         content = replaceHttpsLinks(content);
 
         // Fill in placeholders if content string is empty, truncate it if it's too long
@@ -458,8 +485,8 @@ export default class Toot implements TootObj {
     private repair(): void {
         this.application ??= {name: UNKNOWN};
         this.application.name ??= UNKNOWN;
-        this.language ??= Config.defaultLanguage;
         this.tags.forEach(repairTag);  // Repair Tags
+        this.determineLanguage();      // Repair language
 
         if (this.reblog){
             this.trendingRank ||= this.reblog.trendingRank;
@@ -475,19 +502,19 @@ export default class Toot implements TootObj {
         this.mediaAttachments.forEach((media) => {
             if (media.type == UNKNOWN) {
                 if (isImage(media.remoteUrl)) {
-                    console.info(`Repairing broken image attachment in toot:`, this);
+                    console.info(`${REPAIR_TOOT} Repairing broken image attachment in toot:`, this);
                     media.type = MediaCategory.IMAGE;
                 } else if (isVideo(media.remoteUrl)) {
-                    console.info(`Repairing broken video attachment in toot:`, this);
+                    console.info(`${REPAIR_TOOT} Repairing broken video attachment in toot:`, this);
                     media.type = MediaCategory.VIDEO;
                 } else if (this.uri?.includes(BLUESKY_BRIDGY) && media.previewUrl?.endsWith("/small") && !media.previewRemoteUrl) {
-                    console.info(`Repairing broken bluesky bridge image attachment in toot:`, this);
+                    console.info(`${REPAIR_TOOT} Repairing broken bluesky bridge image attachment in toot:`, this);
                     media.type = MediaCategory.IMAGE;
                 } else {
-                    console.warn(`Unknown media type for URL: '${media.remoteUrl}' for toot:`, this);
+                    console.warn(`${REPAIR_TOOT} Unknown media type for URL: '${media.remoteUrl}' for toot:`, this);
                 }
             } else if (!MEDIA_TYPES.includes(media.type)) {
-                console.warn(`Unknown media of type: '${media.type}' for toot:`, this);
+                console.warn(`${REPAIR_TOOT} Unknown media of type: '${media.type}' for toot:`, this);
             }
         });
 
@@ -497,6 +524,97 @@ export default class Toot implements TootObj {
                 mention.acct += `@${extractDomain(mention.url)}`;
             }
         })
+    }
+
+    // Figure out an appropriate language for the toot based on the content.
+    private determineLanguage(): void {
+        let text = removeMentions(removeEmojis(removeTags(removeLinks(this.contentString()))));
+        text = text.replace(/\s+/g, " ").trim();
+
+        if (text.length < MIN_CHARS_FOR_LANG_DETECT) {
+            this.language ??= Config.defaultLanguage;
+            return;
+        }
+
+        // Use the tinyld language detector to get the detectedLang
+        const detectedLangs = detectAll(text);
+        const detectedLang = detectedLangs[0]?.lang;
+        const detectedLangAccuracy = detectedLangs[0]?.accuracy || 0;
+
+        // Use LanguageDetector to get the altLanguage
+        const altDetectedLangs = LANGUAGE_DETECTOR.detect(text);
+        let altLanguage = altDetectedLangs.length ? altDetectedLangs[0][0] : null;
+        const altLangAccuracy = altDetectedLangs.length ? altDetectedLangs[0][1] : 0;
+        const accuracies = [detectedLangAccuracy, altLangAccuracy];
+        const logObj = {altDetectedLangs, detectedLangs, text, toot: this};
+
+        if (altLanguage && altLanguage in LANGUAGE_CODES) {
+            altLanguage = LANGUAGE_CODES[altLanguage];
+        } else if (altLanguage) {
+            console.warn(`${REPAIR_TOOT} altLanguage "${altLanguage}" found but not in LANGUAGE_CODES!"`, logObj);
+        }
+
+        // If there's nothign detected log a warning (if text is long enough) and set language to default
+        if ((detectedLangs.length + altDetectedLangs.length) == 0) {
+            if (text.length > (MIN_CHARS_FOR_LANG_DETECT * 2)) console.warn(`${REPAIR_TOOT} no language detected`, logObj);
+            this.language ??= Config.defaultLanguage;
+            return;
+        }
+
+        // Return if either language detection matches the toot's language
+        if ((detectedLang && detectedLang == this.language) || (altLanguage && altLanguage == this.language)) {
+            return;
+        }
+
+        let langInfo = `toot.language="${this.language}"`;
+        langInfo +=  `, detectedLang="${detectedLang}" (accuracy: ${detectedLangAccuracy.toPrecision(4)})`;
+        langInfo +=  `, altDetectedLang:="${altLanguage}" (accuracy: ${altLangAccuracy?.toPrecision(4)})`;
+
+        // If both tests agree on the language and the toot has no opinion use that language
+        if (detectedLang && detectedLang == altLanguage && accuracies.some((a) => a > MIN_ACCURACY_FOR_LANG_REPLACE)) {
+            if (this.language && this.language != UNKNOWN) {
+                console.debug(`Both language tests agree: replace "${this.language}" w/ "${detectedLang}". ${langInfo}`, logObj);
+            }
+
+            this.language = detectedLang;
+            return;
+        }
+
+        if (IGNORE_LANGUAGES.includes(detectedLang) || detectedLangAccuracy < MIN_ACCURACY_FOR_LANG_REPLACE)  {
+            if (altLanguage == detectedLang && accuracies.every((a) => a > (MIN_ACCURACY_FOR_LANG_REPLACE / 2))) {
+                // If both tests agree but the accuracy is low, just use the detectedLang
+                console.debug(`Accepting half as accurate language "${detectedLang}" bc both tests agree on it. ${langInfo}`, logObj);
+                this.language = detectedLang;
+            } else if (this.language == LANGUAGE_CODES.english) {
+                // If accuracy is low or ignorable but toot.language is English just use that
+                return;
+            } else if (altLanguage && altLangAccuracy && altLanguage == LANGUAGE_CODES.english && altLangAccuracy > MIN_ACCURACY_FOR_LANG_REPLACE) {
+                // If detectedLang is low accuracy but altLanguage is English, use that
+                console.debug(`Ignoring detected "${detectedLang}" and accepting english from altLanguage. ${langInfo}`, logObj);
+                this.language = LANGUAGE_CODES.english;
+            } else {
+                console.debug(`Ignoring low accuracy or ignorable detected languages. ${langInfo}`, logObj);
+                this.language ??= Config.defaultLanguage;
+            }
+
+            return;
+        }
+
+        if (altLangAccuracy >= MIN_VERY_HIGH_ACCURACY_FOR_LANG_REPLACE) {
+            console.debug(`Accepting high accuracy altLanguage "${altLanguage}" over detectedLang "${detectedLang}". ${langInfo}`, logObj);
+            this.language = altLanguage;
+            return;
+        } else if (detectedLangAccuracy >= MIN_VERY_HIGH_ACCURACY_FOR_LANG_REPLACE && FOREIGN_SCRIPTS.includes(detectedLang)) {
+            console.debug(`Accepting high accuracy detectedLang "${detectedLang}" over altLanguage "${altLanguage}". ${langInfo}`, logObj);
+        } else if (altLanguage && altLanguage != detectedLang) {
+            console.warn(`Want to replace language but detectedLang "${detectedLang}" != altLanguage "${altLanguage}". ${langInfo} `, logObj);
+            this.language ??= Config.defaultLanguage;
+            return;
+        }
+
+        const newLanguage = (detectedLang || altLanguage || Config.defaultLanguage);
+        console.debug(`Replacing existing language with "${newLanguage}"! ${langInfo}`, logObj);
+        this.language = newLanguage;
     }
 
     // Some properties cannot be repaired and/or set until info about the user is available.